feat: Allow specifying node_id
This was a little trickier than I hoped, basically we strip out the
project and event_id info from the nodestore payload before creating
the Event, so at the NodeField level we have no way of knowing these
values.

Instead, I have made it so that a `node_id` key in the data object will
be used as the nodestore id by NodeField.

I also rearranged NodeField and NodeData a little so that the loading
and saving occurs in the same class.